`ASComplete.HandleOverrideCompletion`
